Move the SmallU code out of divisible into small_uint
The divisible module held two unrelated groups of code: the generic
`Divisible` trait with its endianness-aware memcast and mapget helpers, and
the sub-byte machinery that only exists to serve `SmallU`. The second group
belongs beside the type it is about.

Moved to `small_uint`:

- the `bitmask` module, which extracts and inserts sub-byte elements through
  a `Divisible<u8>` interface,
- `SmallUDivisIter`, which wraps a byte iterator and yields `SmallU<N>`,
- the `impl_divisible_bitmask!` macro and its five invocations for `u8`
  through `u128`,
- the `Divisible<SmallU<M>> for SmallU<N>` impls,
- the `SmallU<1>`, `SmallU<2>`, and `SmallU<4>` arms of
  `impl_divisible_self!`, now a separate invocation, and
- the 13 tests that cover the above, in a new `tests` module.

`divisible` keeps the trait, the `memcast` and `mapget` helper modules,
`impl_divisible_memcast!`, `impl_divisible_self!`, and the primitive-only
impls and tests.

The arch backends need no changes: `impl_divisible_bitmask!` stays reachable
at `crate::underlier::impl_divisible_bitmask` through the `pub use
small_uint::*` glob, and the macro body already spells its references with
absolute `$crate::underlier::` paths.
