+++ crates/field/src/underlier/divisible.rs
@@ -303,61 +303,6 @@ pub mod memcast {
 	}
 }
 
-/// Helper functions for Divisible implementations using bitmask operations on sub-byte elements.
-///
-/// These functions work on any type that implements `Divisible<u8>` by extracting
-/// and modifying sub-byte elements through the byte interface.
-pub mod bitmask {
-	use super::{Divisible, SmallU};
-
-	/// Get a sub-byte element at index (LSB-first ordering) without bounds checking.
-	///
-	/// # Safety
-	///
-	/// The caller must ensure that `index < Big::N` (over `SmallU<BITS>` elements).
-	#[inline]
-	pub unsafe fn get<Big, const BITS: usize>(value: &Big, index: usize) -> SmallU<BITS>
-	where
-		Big: Divisible<u8>,
-	{
-		let elems_per_byte = 8 / BITS;
-		let byte_index = index / elems_per_byte;
-		let sub_index = index % elems_per_byte;
-		// Safety: `index < Big::N` over `SmallU<BITS>` implies `byte_index < Big::N` over `u8`.
-		let byte = unsafe { Divisible::<u8>::get_unchecked(value, byte_index) };
-		let shift = sub_index * BITS;
-		SmallU::<BITS>::new(byte >> shift)
-	}
-
-	/// Set a sub-byte element at index (LSB-first ordering), returning modified value, without
-	/// bounds checking.
-	///
-	/// # Safety
-	///
-	/// The caller must ensure that `index < Big::N` (over `SmallU<BITS>` elements).
-	#[inline]
-	pub unsafe fn set<Big, const BITS: usize>(
-		mut value: Big,
-		index: usize,
-		val: SmallU<BITS>,
-	) -> Big
-	where
-		Big: Divisible<u8>,
-	{
-		let elems_per_byte = 8 / BITS;
-		let byte_index = index / elems_per_byte;
-		let sub_index = index % elems_per_byte;
-		// Safety: `index < Big::N` over `SmallU<BITS>` implies `byte_index < Big::N` over `u8`.
-		let byte = unsafe { Divisible::<u8>::get_unchecked(&value, byte_index) };
-		let shift = sub_index * BITS;
-		let mask = (1u8 << BITS) - 1;
-		let new_byte = (byte & !(mask << shift)) | (val.val() << shift);
-		// Safety: `byte_index < Big::N` over `u8`, as above.
-		unsafe { Divisible::<u8>::set_unchecked(&mut value, byte_index, new_byte) };
-		value
-	}
-}
-
 /// Helper functions for Divisible implementations using the get method.
 ///
 /// These functions create iterators by mapping indices through `Divisible::get`,
@@ -394,63 +339,6 @@ pub mod mapget {
 		})
 	}
 }
-
-/// Iterator for dividing an underlier into sub-byte elements (ie. [`SmallU`]).
-///
-/// This iterator wraps a byte iterator and extracts sub-byte elements from each byte.
-/// Generic over the byte iterator type `I`.
-#[derive(Clone)]
-pub struct SmallUDivisIter<I, const N: usize> {
-	byte_iter: I,
-	current_byte: Option<u8>,
-	sub_idx: usize,
-}
-
-impl<I: Iterator<Item = u8>, const N: usize> SmallUDivisIter<I, N> {
-	const ELEMS_PER_BYTE: usize = 8 / N;
-
-	pub fn new(mut byte_iter: I) -> Self {
-		let current_byte = byte_iter.next();
-		Self {
-			byte_iter,
-			current_byte,
-			sub_idx: 0,
-		}
-	}
-}
-
-impl<I: ExactSizeIterator<Item = u8>, const N: usize> Iterator for SmallUDivisIter<I, N> {
-	type Item = SmallU<N>;
-
-	#[inline]
-	fn next(&mut self) -> Option<Self::Item> {
-		let byte = self.current_byte?;
-		let shift = self.sub_idx * N;
-		let result = SmallU::<N>::new(byte >> shift);
-
-		self.sub_idx += 1;
-		if self.sub_idx >= Self::ELEMS_PER_BYTE {
-			self.sub_idx = 0;
-			self.current_byte = self.byte_iter.next();
-		}
-
-		Some(result)
-	}
-
-	#[inline]
-	fn size_hint(&self) -> (usize, Option<usize>) {
-		let remaining_in_current = if self.current_byte.is_some() {
-			Self::ELEMS_PER_BYTE - self.sub_idx
-		} else {
-			0
-		};
-		let remaining_bytes = self.byte_iter.len();
-		let total = remaining_in_current + remaining_bytes * Self::ELEMS_PER_BYTE;
-		(total, Some(total))
-	}
-}
-
-impl<I: ExactSizeIterator<Item = u8>, const N: usize> ExactSizeIterator for SmallUDivisIter<I, N> {}
 
 /// Implements `Divisible` trait using bytemuck memory casting.
 ///
@@ -520,300 +408,11 @@ macro_rules! impl_divisible_memcast {
 #[allow(unused)]
 pub(crate) use impl_divisible_memcast;
 
-/// Implements `Divisible` trait for SmallU types using bitmask operations.
-///
-/// This macro generates `Divisible<SmallU<BITS>>` implementations for a big type
-/// by wrapping byte iteration with bitmasking to extract sub-byte elements.
-macro_rules! impl_divisible_bitmask {
-	// Special case for u8: operates directly on the byte without needing Divisible::<u8>
-	(u8, $($bits:expr),+) => {
-		$(
-			impl $crate::underlier::Divisible<$crate::underlier::SmallU<$bits>> for u8 {
-				const LOG_N: usize = (8usize / $bits).ilog2() as usize;
-
-				#[inline]
-				fn value_iter(value: Self) -> impl ExactSizeIterator<Item = $crate::underlier::SmallU<$bits>> + Send + Clone {
-					$crate::underlier::SmallUDivisIter::new(std::iter::once(value))
-				}
-
-				#[inline]
-				fn ref_iter(value: &Self) -> impl ExactSizeIterator<Item = $crate::underlier::SmallU<$bits>> + Send + Clone + '_ {
-					$crate::underlier::SmallUDivisIter::new(std::iter::once(*value))
-				}
-
-				#[inline]
-				fn slice_iter(slice: &[Self]) -> impl ExactSizeIterator<Item = $crate::underlier::SmallU<$bits>> + Send + Clone + '_ {
-					$crate::underlier::SmallUDivisIter::new(slice.iter().copied())
-				}
-
-				#[inline]
-				unsafe fn get_unchecked(&self, index: usize) -> $crate::underlier::SmallU<$bits> {
-					let shift = index * $bits;
-					$crate::underlier::SmallU::<$bits>::new(*self >> shift)
-				}
-
-				#[inline]
-				unsafe fn set_unchecked(&mut self, index: usize, val: $crate::underlier::SmallU<$bits>) {
-					let shift = index * $bits;
-					let mask = (1u8 << $bits) - 1;
-					*self = (*self & !(mask << shift)) | (val.val() << shift);
-				}
-
-				#[inline]
-				fn broadcast(val: $crate::underlier::SmallU<$bits>) -> Self {
-					if $bits == 1 {
-						// For 1-bit values: 0 -> 0x00, 1 -> 0xFF
-						val.val().wrapping_neg()
-					} else {
-						let mut result = val.val();
-						// Self-replicate to fill the byte
-						let mut current_bits = $bits;
-						while current_bits < 8 {
-							result |= result << current_bits;
-							current_bits *= 2;
-						}
-						result
-					}
-				}
-
-				#[inline]
-				fn from_iter(iter: impl Iterator<Item = $crate::underlier::SmallU<$bits>>) -> Self {
-					const N: usize = 8 / $bits;
-					let mut result: Self = 0;
-					for (i, val) in iter.take(N).enumerate() {
-						$crate::underlier::Divisible::<$crate::underlier::SmallU<$bits>>::set(&mut result, i, val);
-					}
-					result
-				}
-			}
-		)+
-	};
-
-	// General case for types larger than u8: wraps byte iteration
-	($big:ty, $($bits:expr),+) => {
-		$(
-			impl $crate::underlier::Divisible<$crate::underlier::SmallU<$bits>> for $big {
-				const LOG_N: usize = (8 * size_of::<$big>() / $bits).ilog2() as usize;
-
-				#[inline]
-				fn value_iter(value: Self) -> impl ExactSizeIterator<Item = $crate::underlier::SmallU<$bits>> + Send + Clone {
-					$crate::underlier::SmallUDivisIter::new(
-						$crate::underlier::Divisible::<u8>::value_iter(value)
-					)
-				}
-
-				#[inline]
-				fn ref_iter(value: &Self) -> impl ExactSizeIterator<Item = $crate::underlier::SmallU<$bits>> + Send + Clone + '_ {
-					$crate::underlier::SmallUDivisIter::new(
-						$crate::underlier::Divisible::<u8>::ref_iter(value)
-					)
-				}
-
-				#[inline]
-				fn slice_iter(slice: &[Self]) -> impl ExactSizeIterator<Item = $crate::underlier::SmallU<$bits>> + Send + Clone + '_ {
-					$crate::underlier::SmallUDivisIter::new(
-						$crate::underlier::Divisible::<u8>::slice_iter(slice)
-					)
-				}
-
-				#[inline]
-				unsafe fn get_unchecked(&self, index: usize) -> $crate::underlier::SmallU<$bits> {
-					// Safety: the caller guarantees `index < Self::N`.
-					unsafe { $crate::underlier::bitmask::get::<Self, $bits>(self, index) }
-				}
-
-				#[inline]
-				unsafe fn set_unchecked(&mut self, index: usize, val: $crate::underlier::SmallU<$bits>) {
-					// Safety: the caller guarantees `index < Self::N`.
-					*self = unsafe { $crate::underlier::bitmask::set::<Self, $bits>(*self, index, val) };
-				}
-
-				#[inline]
-				fn broadcast(val: $crate::underlier::SmallU<$bits>) -> Self {
-					// First splat to u8, then splat the byte to fill Self
-					let byte = $crate::underlier::Divisible::<$crate::underlier::SmallU<$bits>>::broadcast(val);
-					$crate::underlier::Divisible::<u8>::broadcast(byte)
-				}
-
-				#[inline]
-				fn from_iter(iter: impl Iterator<Item = $crate::underlier::SmallU<$bits>>) -> Self {
-					const N: usize = 8 * size_of::<$big>() / $bits;
-					let mut result: Self = bytemuck::Zeroable::zeroed();
-					for (i, val) in iter.take(N).enumerate() {
-						$crate::underlier::Divisible::<$crate::underlier::SmallU<$bits>>::set(&mut result, i, val);
-					}
-					result
-				}
-			}
-		)+
-	};
-}
-
-#[allow(unused)]
-pub(crate) use impl_divisible_bitmask;
-
-use super::small_uint::SmallU;
-
 // Implement Divisible using memcast for primitive types
 impl_divisible_memcast!(u128, u64, u32, u16, u8);
 impl_divisible_memcast!(u64, u32, u16, u8);
 impl_divisible_memcast!(u32, u16, u8);
 impl_divisible_memcast!(u16, u8);
-
-// Implement Divisible using bitmask for SmallU types
-impl_divisible_bitmask!(u8, 1, 2, 4);
-impl_divisible_bitmask!(u16, 1, 2, 4);
-impl_divisible_bitmask!(u32, 1, 2, 4);
-impl_divisible_bitmask!(u64, 1, 2, 4);
-impl_divisible_bitmask!(u128, 1, 2, 4);
-
-// Divisible for SmallU types that subdivide into smaller SmallU types
-impl Divisible<SmallU<1>> for SmallU<2> {
-	const LOG_N: usize = 1;
-
-	#[inline]
-	fn value_iter(value: Self) -> impl ExactSizeIterator<Item = SmallU<1>> + Send + Clone {
-		mapget::value_iter(value)
-	}
-
-	#[inline]
-	fn ref_iter(value: &Self) -> impl ExactSizeIterator<Item = SmallU<1>> + Send + Clone + '_ {
-		mapget::value_iter(*value)
-	}
-
-	#[inline]
-	fn slice_iter(slice: &[Self]) -> impl ExactSizeIterator<Item = SmallU<1>> + Send + Clone + '_ {
-		mapget::slice_iter(slice)
-	}
-
-	#[inline]
-	unsafe fn get_unchecked(&self, index: usize) -> SmallU<1> {
-		SmallU::<1>::new(self.val() >> index)
-	}
-
-	#[inline]
-	unsafe fn set_unchecked(&mut self, index: usize, val: SmallU<1>) {
-		let mask = 1u8 << index;
-		*self = SmallU::<2>::new((self.val() & !mask) | (val.val() << index));
-	}
-
-	#[inline]
-	fn broadcast(val: SmallU<1>) -> Self {
-		// 0b0 -> 0b00, 0b1 -> 0b11
-		let v = val.val();
-		SmallU::<2>::new(v | (v << 1))
-	}
-
-	#[inline]
-	fn from_iter(iter: impl Iterator<Item = SmallU<1>>) -> Self {
-		iter.chain(std::iter::repeat(SmallU::<1>::new(0)))
-			.take(2)
-			.enumerate()
-			.fold(SmallU::<2>::new(0), |mut acc, (i, val)| {
-				acc.set(i, val);
-				acc
-			})
-	}
-}
-
-impl Divisible<SmallU<1>> for SmallU<4> {
-	const LOG_N: usize = 2;
-
-	#[inline]
-	fn value_iter(value: Self) -> impl ExactSizeIterator<Item = SmallU<1>> + Send + Clone {
-		mapget::value_iter(value)
-	}
-
-	#[inline]
-	fn ref_iter(value: &Self) -> impl ExactSizeIterator<Item = SmallU<1>> + Send + Clone + '_ {
-		mapget::value_iter(*value)
-	}
-
-	#[inline]
-	fn slice_iter(slice: &[Self]) -> impl ExactSizeIterator<Item = SmallU<1>> + Send + Clone + '_ {
-		mapget::slice_iter(slice)
-	}
-
-	#[inline]
-	unsafe fn get_unchecked(&self, index: usize) -> SmallU<1> {
-		SmallU::<1>::new(self.val() >> index)
-	}
-
-	#[inline]
-	unsafe fn set_unchecked(&mut self, index: usize, val: SmallU<1>) {
-		let mask = 1u8 << index;
-		*self = SmallU::<4>::new((self.val() & !mask) | (val.val() << index));
-	}
-
-	#[inline]
-	fn broadcast(val: SmallU<1>) -> Self {
-		// 0b0 -> 0b0000, 0b1 -> 0b1111
-		let mut v = val.val();
-		v |= v << 1;
-		v |= v << 2;
-		SmallU::<4>::new(v)
-	}
-
-	#[inline]
-	fn from_iter(iter: impl Iterator<Item = SmallU<1>>) -> Self {
-		iter.chain(std::iter::repeat(SmallU::<1>::new(0)))
-			.take(4)
-			.enumerate()
-			.fold(SmallU::<4>::new(0), |mut acc, (i, val)| {
-				acc.set(i, val);
-				acc
-			})
-	}
-}
-
-impl Divisible<SmallU<2>> for SmallU<4> {
-	const LOG_N: usize = 1;
-
-	#[inline]
-	fn value_iter(value: Self) -> impl ExactSizeIterator<Item = SmallU<2>> + Send + Clone {
-		mapget::value_iter(value)
-	}
-
-	#[inline]
-	fn ref_iter(value: &Self) -> impl ExactSizeIterator<Item = SmallU<2>> + Send + Clone + '_ {
-		mapget::value_iter(*value)
-	}
-
-	#[inline]
-	fn slice_iter(slice: &[Self]) -> impl ExactSizeIterator<Item = SmallU<2>> + Send + Clone + '_ {
-		mapget::slice_iter(slice)
-	}
-
-	#[inline]
-	unsafe fn get_unchecked(&self, index: usize) -> SmallU<2> {
-		SmallU::<2>::new(self.val() >> (index * 2))
-	}
-
-	#[inline]
-	unsafe fn set_unchecked(&mut self, index: usize, val: SmallU<2>) {
-		let shift = index * 2;
-		let mask = 0b11u8 << shift;
-		*self = SmallU::<4>::new((self.val() & !mask) | (val.val() << shift));
-	}
-
-	#[inline]
-	fn broadcast(val: SmallU<2>) -> Self {
-		// 0bXX -> 0bXXXX
-		let v = val.val();
-		SmallU::<4>::new(v | (v << 2))
-	}
-
-	#[inline]
-	fn from_iter(iter: impl Iterator<Item = SmallU<2>>) -> Self {
-		iter.chain(std::iter::repeat(SmallU::<2>::new(0)))
-			.take(2)
-			.enumerate()
-			.fold(SmallU::<4>::new(0), |mut acc, (i, val)| {
-				acc.set(i, val);
-				acc
-			})
-	}
-}
 
 /// Implements reflexive `Divisible<Self>` for a type (dividing into itself once).
 macro_rules! impl_divisible_self {
@@ -864,124 +463,11 @@ macro_rules! impl_divisible_self {
 #[allow(unused)]
 pub(crate) use impl_divisible_self;
 
-impl_divisible_self!(u8, u16, u32, u64, u128, SmallU<1>, SmallU<2>, SmallU<4>);
+impl_divisible_self!(u8, u16, u32, u64, u128);
 
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::underlier::small_uint::{U1, U2, U4};
-
-	#[test]
-	fn test_divisible_u8_u4() {
-		let val: u8 = 0x34;
-
-		// Test get - LSB first: nibbles
-		assert_eq!(Divisible::<U4>::get(&val, 0), U4::new(0x4));
-		assert_eq!(Divisible::<U4>::get(&val, 1), U4::new(0x3));
-
-		// Test set
-		let mut modified = val;
-		Divisible::<U4>::set(&mut modified, 0, U4::new(0xF));
-		assert_eq!(modified, 0x3F);
-		let mut modified = val;
-		Divisible::<U4>::set(&mut modified, 1, U4::new(0xA));
-		assert_eq!(modified, 0xA4);
-
-		// Test ref_iter
-		let parts: Vec<U4> = Divisible::<U4>::ref_iter(&val).collect();
-		assert_eq!(parts.len(), 2);
-		assert_eq!(parts[0], U4::new(0x4));
-		assert_eq!(parts[1], U4::new(0x3));
-
-		// Test value_iter
-		let parts: Vec<U4> = Divisible::<U4>::value_iter(val).collect();
-		assert_eq!(parts.len(), 2);
-		assert_eq!(parts[0], U4::new(0x4));
-		assert_eq!(parts[1], U4::new(0x3));
-
-		// Test slice_iter
-		let vals = [0x34u8, 0x56u8];
-		let parts: Vec<U4> = Divisible::<U4>::slice_iter(&vals).collect();
-		assert_eq!(parts.len(), 4);
-		assert_eq!(parts[0], U4::new(0x4));
-		assert_eq!(parts[1], U4::new(0x3));
-		assert_eq!(parts[2], U4::new(0x6));
-		assert_eq!(parts[3], U4::new(0x5));
-	}
-
-	#[test]
-	fn test_divisible_u16_u4() {
-		let val: u16 = 0x1234;
-
-		// Test get - LSB first: nibbles
-		assert_eq!(Divisible::<U4>::get(&val, 0), U4::new(0x4));
-		assert_eq!(Divisible::<U4>::get(&val, 1), U4::new(0x3));
-		assert_eq!(Divisible::<U4>::get(&val, 2), U4::new(0x2));
-		assert_eq!(Divisible::<U4>::get(&val, 3), U4::new(0x1));
-
-		// Test set
-		let mut modified = val;
-		Divisible::<U4>::set(&mut modified, 1, U4::new(0xF));
-		assert_eq!(modified, 0x12F4);
-
-		// Test ref_iter
-		let parts: Vec<U4> = Divisible::<U4>::ref_iter(&val).collect();
-		assert_eq!(parts.len(), 4);
-		assert_eq!(parts[0], U4::new(0x4));
-		assert_eq!(parts[3], U4::new(0x1));
-	}
-
-	#[test]
-	fn test_divisible_u16_u2() {
-		// 0b1011_0010_1101_0011 = 0xB2D3
-		let val: u16 = 0b1011001011010011;
-
-		// Test get - LSB first: 2-bit chunks
-		assert_eq!(Divisible::<U2>::get(&val, 0), U2::new(0b11)); // bits 0-1
-		assert_eq!(Divisible::<U2>::get(&val, 1), U2::new(0b00)); // bits 2-3
-		assert_eq!(Divisible::<U2>::get(&val, 7), U2::new(0b10)); // bits 14-15
-
-		// Test ref_iter
-		let parts: Vec<U2> = Divisible::<U2>::ref_iter(&val).collect();
-		assert_eq!(parts.len(), 8);
-		assert_eq!(parts[0], U2::new(0b11));
-		assert_eq!(parts[7], U2::new(0b10));
-	}
-
-	#[test]
-	fn test_divisible_u16_u1() {
-		// 0b1010_1100_0011_0101 = 0xAC35
-		let val: u16 = 0b1010110000110101;
-
-		// Test get - LSB first: individual bits
-		assert_eq!(Divisible::<U1>::get(&val, 0), U1::new(1)); // bit 0
-		assert_eq!(Divisible::<U1>::get(&val, 1), U1::new(0)); // bit 1
-		assert_eq!(Divisible::<U1>::get(&val, 15), U1::new(1)); // bit 15
-
-		// Test set
-		let mut modified = val;
-		Divisible::<U1>::set(&mut modified, 0, U1::new(0));
-		assert_eq!(modified, 0b1010110000110100);
-
-		// Test ref_iter
-		let parts: Vec<U1> = Divisible::<U1>::ref_iter(&val).collect();
-		assert_eq!(parts.len(), 16);
-		assert_eq!(parts[0], U1::new(1));
-		assert_eq!(parts[15], U1::new(1));
-	}
-
-	#[test]
-	fn test_divisible_u64_u4() {
-		let val: u64 = 0x123456789ABCDEF0;
-
-		// Test get - LSB first: nibbles
-		assert_eq!(Divisible::<U4>::get(&val, 0), U4::new(0x0));
-		assert_eq!(Divisible::<U4>::get(&val, 1), U4::new(0xF));
-		assert_eq!(Divisible::<U4>::get(&val, 15), U4::new(0x1));
-
-		// Iterating a u64 as nibbles yields 64 / 4 = 16 parts.
-		assert_eq!(Divisible::<U4>::ref_iter(&val).count(), 16);
-	}
 
 	#[test]
 	fn test_divisible_u32_u8_slice() {
@@ -1017,56 +503,6 @@ mod tests {
 	fn test_broadcast_u128_u32() {
 		let result: u128 = Divisible::<u32>::broadcast(0xDEADBEEF);
 		assert_eq!(result, 0xDEADBEEFDEADBEEFDEADBEEFDEADBEEF);
-	}
-
-	#[test]
-	fn test_broadcast_u8_u4() {
-		let result: u8 = Divisible::<U4>::broadcast(U4::new(0x5));
-		assert_eq!(result, 0x55);
-	}
-
-	#[test]
-	fn test_broadcast_u16_u4() {
-		let result: u16 = Divisible::<U4>::broadcast(U4::new(0xA));
-		assert_eq!(result, 0xAAAA);
-	}
-
-	#[test]
-	fn test_broadcast_u8_u2() {
-		let result: u8 = Divisible::<U2>::broadcast(U2::new(0b11));
-		assert_eq!(result, 0xFF);
-		let result: u8 = Divisible::<U2>::broadcast(U2::new(0b01));
-		assert_eq!(result, 0x55);
-	}
-
-	#[test]
-	fn test_broadcast_u8_u1() {
-		let result: u8 = Divisible::<U1>::broadcast(U1::new(0));
-		assert_eq!(result, 0x00);
-		let result: u8 = Divisible::<U1>::broadcast(U1::new(1));
-		assert_eq!(result, 0xFF);
-	}
-
-	#[test]
-	fn test_broadcast_smallu2_from_smallu1() {
-		let result: SmallU<2> = Divisible::<SmallU<1>>::broadcast(SmallU::<1>::new(0));
-		assert_eq!(result.val(), 0b00);
-		let result: SmallU<2> = Divisible::<SmallU<1>>::broadcast(SmallU::<1>::new(1));
-		assert_eq!(result.val(), 0b11);
-	}
-
-	#[test]
-	fn test_broadcast_smallu4_from_smallu1() {
-		let result: SmallU<4> = Divisible::<SmallU<1>>::broadcast(SmallU::<1>::new(0));
-		assert_eq!(result.val(), 0b0000);
-		let result: SmallU<4> = Divisible::<SmallU<1>>::broadcast(SmallU::<1>::new(1));
-		assert_eq!(result.val(), 0b1111);
-	}
-
-	#[test]
-	fn test_broadcast_smallu4_from_smallu2() {
-		let result: SmallU<4> = Divisible::<SmallU<2>>::broadcast(SmallU::<2>::new(0b10));
-		assert_eq!(result.val(), 0b1010);
 	}
 
 	#[test]
@@ -1107,11 +543,5 @@ mod tests {
 		let result: u64 = Divisible::<u16>::from_iter([0x1234, 0x5678, 0x9ABC].into_iter());
 		// Only 3 elements provided, 4th should be 0
 		assert_eq!(result, 0x0000_9ABC_5678_1234);
-	}
-
-	#[test]
-	fn test_from_iter_smallu() {
-		let result: u8 = Divisible::<U4>::from_iter([U4::new(0xA), U4::new(0xB)].into_iter());
-		assert_eq!(result, 0xBA);
 	}
 }
